LayoutTests/editing/selection/selection-setSelectionRange-frameselection-expected.txt



PASS setSelectionRange() should not update FrameSelection if the target element is not focused.
PASS setRangeText() should not update FrameSelection if the target element is not focused.
